Revert "showing collage modal every time, About Page"
This reverts commit 2063a01668301c5d22bed8f10e5c297a7435fcd0.

+++ app/actions/collage-actions/collage-actions.js
@@ -10,7 +10,7 @@ export function setCollageMode({ isCollageMode }) {
 
         const alreadySeenModalMessage = LocalStorageService.getItem('seenCollageMessage');
 
-        if (isCollageMode) {
+        if (isCollageMode && !alreadySeenModalMessage ) {
             dispatch(showNotification({ message: 'מצב קולאז׳ מופעל, בחרו עד 4 ממים לשילוב יחדיו' }));
             LocalStorageService.setItem('seenCollageMessage', 'true')
         }
